pyproject.toml
[project]
name = "ai_voice_assistant"
version = "0.2.0"
description = "AI Voice Assistant"
requires-python = ">=3.12"

authors = ["roktober <vkvc916@gmail.com>"]
readme = "README.md"

dependencies = [
    "elevenlabs",
    "openai",
    "python-dotenv",
    "pydantic-settings",
    "python-telegram-bot",
    "tenacity",
    "black",
    "ruff",
    "mypy",
    "isort",
    "google-genai"
]


[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.isort]
line_length = 120
profile = "black"

[tool.mypy]
plugins = "pydantic.mypy"

strict = "true"
warn_return_any = "true"
warn_unreachable = "true"
implicit_optional = "false"
disallow_incomplete_defs = "true"
disallow_untyped_defs = "true"
disallow_untyped_decorators = "true"
disallow_any_unimported = "false"
disallow_any_expr = "false"
disallow_any_decorated = "true"
disallow_any_explicit = "true"
disallow_any_generics = "true"
disallow_subclassing_any = "true"
ignore_missing_imports = "true"
